+++ src/lib/crm/types.ts
